Remove unwanted DAO layer for DCR
